(PC-29971)[PRO] refactor: Deleting GetOfferersAdapter Adapter.
